Major Refactors to app package, mqtt package (#45)
This change is mostly a cleanup change and refactors the application to be more "go canoical" as I tend to like it. Structs and interfaces are named better, and packages behave and are organized a bit better. This addresses many of the issues in #41 specifically in the app, mqtt, and handlers packages.